MTCNN/requirements.txt
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.21.0
tqdm>=4.65.0
matplotlib>=3.7.0
tensorboard>=2.13.0
torch==2.5.1+cu118 torchvision==0.20.1+cu118 torchaudio==2.5.1+cu118 --index-url https://download.pytorch.org/whl/cu118